Removing blake2 and only using siphash.
The primary tradeoff here is: does siphash have sufficient preimage resistance as to require random, brute-force searches for each side of the equation. This is guaranteed with a cryptographic hash and can be unclear otherwise.

The main motivator to change to siphash is for its superior performance and simplicity. Computing the hash is done extensively in the main inner loop. Having a computationally slow hash function means more time overhead is spent hashing compared to memory accesses.

The reason to not leave it parameterized in code is for simplicity or potential lost optimizations from dynamic-dispatch or inlining.